+++ src/encoded/tests/test_metadata_tsv_workbook.py
@@ -118,7 +118,7 @@ class TestMetadataTSVWorkbook:
         TestMetadataTSVHelper.check_key_and_length(header1, 'Metadata TSV Download')
         TestMetadataTSVHelper.check_key_and_length(header2, 'Suggested command to download: ')
         TestMetadataTSVHelper.check_key_and_length(header3, 'FileDownloadURL')
-        assert len(parsed[3:]) == 20  # there are 20 entries in the workbook right now, including extra files
+        assert len(parsed[3:]) == 21  # there are 21 entries in the workbook right now, including extra files
         # test for various types
         TestMetadataTSVHelper.check_type_length(es_testapp, 'AlignedReads', 3)
         TestMetadataTSVHelper.check_type_length(es_testapp, 'UnalignedReads', 5)
@@ -126,6 +126,8 @@ class TestMetadataTSVWorkbook:
         TestMetadataTSVHelper.check_type_length(es_testapp, 'ReferenceFile', 2)
         TestMetadataTSVHelper.check_type_length(es_testapp, 'OutputFile', 2)
         TestMetadataTSVHelper.check_type_length(es_testapp, 'SupplementaryFile', 2)
+        TestMetadataTSVHelper.check_type_length(es_testapp, 'HistologyImage', 1)
+
 
         res = es_testapp.post_json('/metadata/', {'type': 'OutputFile', 'include_extra_files': True})
         tsv = res._app_iter[0]